refactor(core): extract MODEL_KIND constants to constants.ts, add coverage
Moves MODEL_KIND/MODEL_KIND_TEXT_MAP out of list.ts into a dedicated
constants.ts, matching the deployment/shared.ts precedent and the
file-directory-structure convention of feature-root constants files.
Adds unit coverage for MODEL_KIND_TEXT_MAP, closing the test gap that
let the numeric-vs-string key mismatch ship in 97517441 undetected.
